# Speed up interval rounding