@nik9000 2020-10-05
I'm running performance tests locally to see what I see on some real data.