@nik9000 2020-10-05
The data that I have to run performance tests for this actually doesn't work well because it contains outliers which disable this optimization - there are some docs with dates in 1970 and some on 2050 but the bulk of them are around 2015 or something. That kind of thing defeats this optimization. So we're going to have to rely on the lower level performance test here. I'll have a think about outliers and how we can stop them from defeating the optimization.